+++ api/src/main/java/org/allaymc/api/entity/component/EntityPlayerClientComponent.java
@@ -2,6 +2,7 @@ package org.allaymc.api.entity.component;
 
 import org.allaymc.api.entity.interfaces.EntityPlayer;
 import org.allaymc.api.message.MayContainTrKey;
+import org.allaymc.api.message.MessageReceiver;
 import org.allaymc.api.message.TrKeys;
 import org.allaymc.api.player.ClientState;
 import org.allaymc.api.player.LoginData;
@@ -16,7 +17,12 @@ import java.util.Collection;
  *
  * @author daoge_cmd
  */
-public interface EntityPlayerClientComponent extends EntityComponent {
+public interface EntityPlayerClientComponent extends EntityComponent, MessageReceiver {
+
+    Speed DEFAULT_SPEED = new Speed(0.1, 1.0);
+    Speed DEFAULT_FLY_SPEED = new Speed(0.05, 1.0);
+    Speed DEFAULT_VERTICAL_FLY_SPEED = new Speed(1.0, 1.0);
+
     /**
      * Get the login data of the client.
      *
@@ -196,6 +202,111 @@ public interface EntityPlayerClientComponent extends EntityComponent {
      * Clear the title that is being displayed on the player's screen
      */
     void clearTitle();
+
+    /**
+     * Sends the experience level to the client.
+     *
+     * @param value the experience level to be sent.
+     */
+    void sendExperienceLevel(int value);
+
+    /**
+     * Sends the experience progress to the client.
+     *
+     * @param value the progress value to be sent.
+     */
+    void sendExperienceProgress(float value);
+
+    /**
+     * Sends the food level to the client.
+     *
+     * @param value the food level to be sent.
+     */
+    void sendFoodLevel(int value);
+
+    /**
+     * Sends the food saturation level to the client.
+     *
+     * @param value the food saturation level to send.
+     */
+    void sendFoodSaturationLevel(float value);
+
+    /**
+     * Sends the food exhaustion level to the client.
+     *
+     * @param value the exhaustion level to be sent.
+     */
+    void sendFoodExhaustionLevel(float value);
+
+    /**
+     * Represents a speed value controlled by a base speed and a multiplier. The actual
+     * speed value is the multiplication of the base speed with the multiplier.
+     *
+     * @param baseSpeed  the base speed value
+     * @param multiplier the multiplier of the base speed value
+     */
+    record Speed(double baseSpeed, double multiplier) {
+        /**
+         * Calculates the actual speed by multiplying the base speed with the multiplier.
+         *
+         * @return the calculated speed as a product of the base speed and multiplier
+         */
+        public double calculate() {
+            return baseSpeed * multiplier;
+        }
+
+        /**
+         * Creates a new {@code Speed} instance by adding the specified multiplier to the current multiplier.
+         *
+         * @param multiplier the multiplier to be added to the current multiplier
+         * @return a new {@code Speed} instance with the updated multiplier
+         */
+        public Speed addMultiplier(double multiplier) {
+            return new Speed(this.baseSpeed, this.multiplier + multiplier);
+        }
+    }
+
+    /**
+     * Get the speed of the player.
+     *
+     * @return The speed of the player
+     */
+    Speed getSpeed();
+
+    /**
+     * Set the speed of the player.
+     *
+     * @param speed The speed to set
+     */
+    void setSpeed(Speed speed);
+
+    /**
+     * Get the fly speed of the player.
+     *
+     * @return The fly speed of the player
+     */
+    Speed getFlySpeed();
+
+    /**
+     * Sets the fly speed of the player.
+     *
+     * @param flySpeed the fly speed to set
+     */
+    void setFlySpeed(Speed flySpeed);
+
+    /**
+     * Get the vertical fly speed of the player.
+     *
+     * @return The vertical fly speed of the player
+     */
+    Speed getVerticalFlySpeed();
+
+    /**
+     * Sets the vertical fly speed of the player.
+     *
+     * @param verticalFlySpeed the vertical fly speed to set
+     */
+    void setVerticalFlySpeed(Speed verticalFlySpeed);
 
     /**
      * Sends a packet to the client.